crmai/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# PostgreSQL support
psycopg2-binary==2.9.9

# SuperTokens integration
supertokens-python==0.21.0

# File upload and storage
boto3==1.35.0
Pillow==10.4.0

# HTTP requests
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Rate limiting
Flask-Limiter==3.8.0

# Date/time handling
python-dateutil==2.9.0

# JSON Web Tokens
PyJWT==2.9.0

# Validation
marshmallow==3.22.0

# Background tasks (optional)
celery==5.4.0
redis==5.1.1

# WSGI Server (Required for production)
gunicorn==22.0.0
gevent==24.2.1

# Additional missing dependencies
python-multipart==0.0.9

# Testing dependencies
pytest==8.3.0
pytest-flask==1.3.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==33.1.0
pytest-mock==3.12.0
pytest-asyncio==0.25.0
